pyproject.toml
[tool.poetry]
name = "booktracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "booktracker" }
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"
djoser = "^2.2.3"
python-dotenv = "^1.0.1"
django-extensions = "^3.2.3"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
flake8 = "^7.1.1"
django-filter = "^24.3"
elasticsearch-dsl = "^8.16.0"
django-elasticsearch-dsl = "^8.0"
celery = {extras = ["redis"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"